+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:blurting/config/app_config.dart';
 import 'package:blurting/token.dart';
-import 'package:blurting/startpage/startpage.dart';
+import 'package:blurting/startpage/start_page.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/mainapp.dart';
 import 'package:flutter/services.dart';
