+++ lib/pages/policy/policyTwo.dart
@@ -1,6 +1,6 @@
 import 'package:blurting/mainapp.dart';
 import 'package:blurting/pages/useguide/useguidepagetwo.dart';
-import 'package:blurting/signup_questions/phonenumber.dart';
+import 'package:blurting/signup_questions/phone_number.dart';
 import 'package:blurting/token.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/config/app_config.dart';
