+++ lib/pages/useguide/useguide_page_four.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:blurting/pages/useGuide/useguidepageseven.dart';
+import 'package:blurting/pages/useguide/useguide_page_five.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/Utils/provider.dart';
 
@@ -12,19 +12,20 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: UseGuidePageSix(),
+      home: UseGuidePageFour(),
     );
   }
 }
 
-class UseGuidePageSix extends StatefulWidget {
+class UseGuidePageFour extends StatefulWidget {
   @override
-  _UseGuidePageSixState createState() => _UseGuidePageSixState();
+  _UseGuidePageFourState createState() => _UseGuidePageFourState();
 }
 
-class _UseGuidePageSixState extends State<UseGuidePageSix>
+class _UseGuidePageFourState extends State<UseGuidePageFour>
     with TickerProviderStateMixin {
   AnimationController? _animationController;
+
   Animation<double>? _progressAnimation;
   bool isVisible = true;
   Timer? _blinkTimer;
@@ -35,7 +36,7 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
         .push(
       PageRouteBuilder(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            UseGuidePageSeven(),
+            UseGuidePageFive(),
         transitionsBuilder: (context, animation, secondaryAnimation, child) {
           return FadeTransition(opacity: animation, child: child);
         },
@@ -55,8 +56,8 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
     );
 
     _progressAnimation = Tween<double>(
-      begin: 5 / 7, // 시작 너비 (30%)
-      end: 6 / 7, // 종료 너비 (40%)
+      begin: 3 / 7, // 시작 너비 (30%)
+      end: 4 / 7, // 종료 너비 (40%)
     ).animate(
         CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
       ..addListener(() {
@@ -119,14 +120,14 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
                         color: mainColor.pink),
                     children: [
                       TextSpan(
-                        text: '귓속말',
+                        text: '블러팅',
                       ),
                       TextSpan(
                         text: '에서',
                         style: TextStyle(
                           fontSize: 20,
                           fontWeight: FontWeight.w500,
-                        ),
+                        ), // 원하는 색으로 변경하세요.
                       ),
                     ],
                   ),
@@ -134,7 +135,7 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
               ),
               Container(
                 alignment: Alignment.centerLeft,
-                child: Text("상대방의 프로필을 누르면 상대방의",
+                child: Text("상대방의 프로필을 누르면 귓속말을 걸 수",
                     style: TextStyle(
                       fontSize: 18,
                       fontWeight: FontWeight.w500,
@@ -144,7 +145,7 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
               ),
               Container(
                 alignment: Alignment.centerLeft,
-                child: Text("프로필 카드를 볼 수 있어요!",
+                child: Text("있어요!",
                     style: TextStyle(
                       fontSize: 18,
                       fontWeight: FontWeight.w500,
@@ -152,29 +153,39 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
                       fontFamily: 'Pretendard',
                     )),
               ),
-              SizedBox(height: 100),
-              Column(children: <Widget>[
-                Stack(
-                  clipBehavior: Clip.none, // 화면 밑에 짤리는 부분 나오게 하기
-                  children: <Widget>[
-                    Container(
-                      width: 350,
-                      height: 64,
-                      child: Image.asset("assets/images/useguidesix.png"),
+              SizedBox(height: 40),
+              Stack(
+                clipBehavior: Clip.none, // 화면 밑에 짤리는 부분 나오게 하기
+                children: <Widget>[
+                  Align(
+                    alignment: Alignment.center,
+                    child: Container(
+                      width: 294,
+                      height: 338,
+                      child: Image.asset("assets/images/profilecard.png"),
                     ),
-                    Positioned(
-                      left: 50,
-                      top: 50,
-                      child: Image.asset(
-                        "assets/images/pointer.png",
-                        width: 36.7,
-                        height: 47,
-                      ),
+                  ),
+                  Positioned(
+                    left: 150, // 원하는 위치로 조정하세요.
+                    top: 270, // 원하는 위치로 조정하세요.
+                    child: Image.asset(
+                      "assets/images/pointer.png",
+                      width: 36.7,
+                      height: 47,
                     ),
-                  ],
-                )
-              ]),
-                            Expanded(
+                  ),
+                  Positioned(
+                    left: 130, // 원하는 위치로 조정하세요.
+                    top: 270, // 원하는 위치로 조정하세요.
+                    child: Image.asset(
+                      "assets/images/pointer.png",
+                      width: 36.7,
+                      height: 47,
+                    ),
+                  ),
+                ],
+              ),
+              Expanded(
                 child: Container(
                   alignment: Alignment.bottomCenter,
                   padding: EdgeInsets.only(bottom: 40),
@@ -194,6 +205,7 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
             ],
           ),
         ),
+
         floatingActionButton: Padding(
           padding: const EdgeInsets.fromLTRB(20, 0, 20, 80), // 좌우 마진을 16.0으로 설정
           child: Stack(
@@ -212,7 +224,7 @@ class _UseGuidePageSixState extends State<UseGuidePageSix>
                         (_progressAnimation?.value ?? 0.3) -
                     32, // 좌우 패딩을 고려하여 너비 조정
                 decoration: BoxDecoration(
-                  color: mainColor.pink, // 다크핑크 색상을 사용자 지정 색상으로 가정
+                  color: mainColor.pink,
                   borderRadius: BorderRadius.circular(4.0),
                 ),
               ),
