+++ lib/pages/useguide/useguide_page_six.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:blurting/pages/useguide/useguidepagefive.dart';
+import 'package:blurting/pages/useGuide/useguide_page_seven.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/Utils/provider.dart';
 
@@ -12,20 +12,19 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: UseGuidePageFour(),
+      home: UseGuidePageSix(),
     );
   }
 }
 
-class UseGuidePageFour extends StatefulWidget {
+class UseGuidePageSix extends StatefulWidget {
   @override
-  _UseGuidePageFourState createState() => _UseGuidePageFourState();
+  _UseGuidePageSixState createState() => _UseGuidePageSixState();
 }
 
-class _UseGuidePageFourState extends State<UseGuidePageFour>
+class _UseGuidePageSixState extends State<UseGuidePageSix>
     with TickerProviderStateMixin {
   AnimationController? _animationController;
-
   Animation<double>? _progressAnimation;
   bool isVisible = true;
   Timer? _blinkTimer;
@@ -36,7 +35,7 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
         .push(
       PageRouteBuilder(
         pageBuilder: (context, animation, secondaryAnimation) =>
-            UseGuidePageFive(),
+            UseGuidePageSeven(),
         transitionsBuilder: (context, animation, secondaryAnimation, child) {
           return FadeTransition(opacity: animation, child: child);
         },
@@ -56,8 +55,8 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
     );
 
     _progressAnimation = Tween<double>(
-      begin: 3 / 7, // 시작 너비 (30%)
-      end: 4 / 7, // 종료 너비 (40%)
+      begin: 5 / 7, // 시작 너비 (30%)
+      end: 6 / 7, // 종료 너비 (40%)
     ).animate(
         CurvedAnimation(parent: _animationController!, curve: Curves.easeInOut))
       ..addListener(() {
@@ -120,14 +119,14 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
                         color: mainColor.pink),
                     children: [
                       TextSpan(
-                        text: '블러팅',
+                        text: '귓속말',
                       ),
                       TextSpan(
                         text: '에서',
                         style: TextStyle(
                           fontSize: 20,
                           fontWeight: FontWeight.w500,
-                        ), // 원하는 색으로 변경하세요.
+                        ),
                       ),
                     ],
                   ),
@@ -135,7 +134,7 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
               ),
               Container(
                 alignment: Alignment.centerLeft,
-                child: Text("상대방의 프로필을 누르면 귓속말을 걸 수",
+                child: Text("상대방의 프로필을 누르면 상대방의",
                     style: TextStyle(
                       fontSize: 18,
                       fontWeight: FontWeight.w500,
@@ -145,7 +144,7 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
               ),
               Container(
                 alignment: Alignment.centerLeft,
-                child: Text("있어요!",
+                child: Text("프로필 카드를 볼 수 있어요!",
                     style: TextStyle(
                       fontSize: 18,
                       fontWeight: FontWeight.w500,
@@ -153,38 +152,28 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
                       fontFamily: 'Pretendard',
                     )),
               ),
-              SizedBox(height: 40),
-              Stack(
-                clipBehavior: Clip.none, // 화면 밑에 짤리는 부분 나오게 하기
-                children: <Widget>[
-                  Align(
-                    alignment: Alignment.center,
-                    child: Container(
-                      width: 294,
-                      height: 338,
-                      child: Image.asset("assets/images/profilecard.png"),
+              SizedBox(height: 100),
+              Column(children: <Widget>[
+                Stack(
+                  clipBehavior: Clip.none, // 화면 밑에 짤리는 부분 나오게 하기
+                  children: <Widget>[
+                    Container(
+                      width: 350,
+                      height: 64,
+                      child: Image.asset("assets/images/useguidesix.png"),
                     ),
-                  ),
-                  Positioned(
-                    left: 150, // 원하는 위치로 조정하세요.
-                    top: 270, // 원하는 위치로 조정하세요.
-                    child: Image.asset(
-                      "assets/images/pointer.png",
-                      width: 36.7,
-                      height: 47,
+                    Positioned(
+                      left: 50,
+                      top: 50,
+                      child: Image.asset(
+                        "assets/images/pointer.png",
+                        width: 36.7,
+                        height: 47,
+                      ),
                     ),
-                  ),
-                  Positioned(
-                    left: 130, // 원하는 위치로 조정하세요.
-                    top: 270, // 원하는 위치로 조정하세요.
-                    child: Image.asset(
-                      "assets/images/pointer.png",
-                      width: 36.7,
-                      height: 47,
-                    ),
-                  ),
-                ],
-              ),
+                  ],
+                )
+              ]),
               Expanded(
                 child: Container(
                   alignment: Alignment.bottomCenter,
@@ -205,7 +194,6 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
             ],
           ),
         ),
-
         floatingActionButton: Padding(
           padding: const EdgeInsets.fromLTRB(20, 0, 20, 80), // 좌우 마진을 16.0으로 설정
           child: Stack(
@@ -224,7 +212,7 @@ class _UseGuidePageFourState extends State<UseGuidePageFour>
                         (_progressAnimation?.value ?? 0.3) -
                     32, // 좌우 패딩을 고려하여 너비 조정
                 decoration: BoxDecoration(
-                  color: mainColor.pink,
+                  color: mainColor.pink, // 다크핑크 색상을 사용자 지정 색상으로 가정
                   borderRadius: BorderRadius.circular(4.0),
                 ),
               ),
