+++ lib/pages/useguide/useguide_page_three.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:blurting/pages/useguide/useguidepagefour.dart';
+import 'package:blurting/pages/useguide/useguide_page_four.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/Utils/provider.dart';
